Goals + AutoRespawn
What it says on the tin, there are goals that the meatball can hit that destroy it and add to a score, and it then goes to re-serve the ball. Also there are floors that if the meatball hits it, it halts it and re-serves it. These require the tag "Respawn."